@@ -4,7 +4,7 @@
   "description": "make sense of a photo archive",
   "main": "src/index.js",
   "scripts": {
-    "start": "node src/index.js",
+    "start": "ts-node src/index.ts",
     "lint": "eslint src",
     "test": "tape test/**/*.js | tap-spec"
   },
@@ -24,34 +24,44 @@
     "test"
   ],
   "dependencies": {
-    "aws-amplify": "0.2.11",
-    "aws-sdk": "^2.325.0",
+    "aws-amplify": "1.2.2",
+    "aws-sdk": "^2.562.0",
     "babel-register": "^6.26.0",
-    "colors": "^1.1.2",
+    "colors": "^1.4.0",
     "exif": "^0.6.0",
-    "express": "^4.16.2",
+    "express": "^4.17.1",
     "generate": "^0.14.0",
-    "glob": "^7.1.2",
-    "image-size": "^0.6.1",
+    "glob": "^7.1.5",
+    "image-size": "^0.8.3",
     "isomorphic-fetch": "^2.2.1",
     "kairos-api": "^0.1.3",
-    "node-base64-image": "^1.0.4",
-    "omit-empty": "^0.4.1",
+    "node-base64-image": "^1.0.6",
+    "omit-empty": "^1.0.0",
     "pre-commit": "^1.2.2",
-    "pug": "^2.0.0-rc.4",
+    "pug": "^2.0.4",
     "retry": "^0.12.0",
-    "sharp": "^0.18.4",
-    "yargs": "^12.0.2"
+    "sharp": "^0.23.2",
+    "yargs": "^14.2.0"
   },
   "devDependencies": {
+    "@types/aws-lambda": "^8.10.34",
+    "@types/dotenv": "^8.2.0",
+    "@types/node": "^12.12.5",
+    "@types/tape": "^4.2.33",
+    "@typescript-eslint/eslint-plugin": "^2.6.0",
+    "@typescript-eslint/parser": "^2.6.0",
+    "amazon-cognito-identity-js": "^3.2.0",
+    "aws-creds-loader": "^0.1.0",
     "babel-core": "^6.26.3",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-3": "^6.24.1",
-    "eslint": "^4.9.0",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.8.0",
-    "tap-spec": "^4.1.1",
-    "tape": "^4.8.0"
+    "eslint": "^6.6.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.18.2",
+    "tap-spec": "^5.0.0",
+    "tape": "^4.11.0",
+    "ts-node": "^8.4.1",
+    "typescript": "^3.6.4"
   }
 }
